nyduds: move filesystem related code into fs_service.rs
Move filesystem specific code into fs_service.rs.
